Replace keygiftsdoor with more general 3 phase env
ThreePhaseDelayedReward in delayed_reward_multiphase.py.  Always starts with KeyEnv in first phase, but distractor and last phases can be specified by user
