+++ gym_minigrid/envs/__init__.py
@@ -24,6 +24,6 @@ from gym_minigrid.envs.distshift import *
 from gym_minigrid.envs.doorkeyoptional import *
 from gym_minigrid.envs.opengifts import *
 from gym_minigrid.envs.fetchkey import *
-from gym_minigrid.envs.key_to_gifts_to_door import *
 from gym_minigrid.envs.keygoal import *
 from gym_minigrid.envs.goalkeyoptional import *
+from gym_minigrid.envs.delayed_reward_multiphase import *
